feat(server): optimisticallyApplyActionOnAllFlatEntityMaps facto
